refactor: move match components into matches directory
