+++ app/components/Matches/GameInfo.jsx
@@ -1,22 +1,21 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import champList from '../../seed';
-import styles from '../styles/Match.css';
+import champList from '../../../seed';
 
-const Match = (props) => {
+const GameInfo = (props) => {
   const champ = champList.Heroes[props.champion].key;
   const datePlayed = JSON.stringify(new Date(props.timestamp));
   return (
-    <div className={styles.match_container}>
+    <div >
       <h3>{datePlayed}</h3>
       <img src={`/images/champions/${champ}.png`} alt={`${champ} icon`} />
     </div>
   );
 };
 
-Match.propTypes = {
+GameInfo.propTypes = {
   champion: PropTypes.number.isRequired,
   timestamp: PropTypes.number.isRequired,
 };
 
-export default Match;
+export default GameInfo;
