+++ app/components/Matches/Match.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import styles from '../../styles/Match.css';
+import GameInfo from './GameInfo';
+
+const Match = () => (
+  <div className={styles.match_container}>
+    <GameInfo />
+  </div>
+);
+
+export default Match;
