+++ app/components/index.jsx
@@ -7,7 +7,7 @@ export { default as Leagues } from './Leagues';
 export { default as Navbar } from './Navbar';
 export { default as NotFound } from './NotFound';
 export { default as Masteries } from './Masteries';
-export { default as Match } from './Match';
+export { default as Match } from './Matches/Match';
 export { default as ProfileSummary } from './ProfileSummary';
 export { default as Runes } from './Runes';
 export { default as QueueRank } from './QueueRank';
