+++ app/styles/Match.css
@@ -1,3 +1,12 @@
 .match_container {
   width: 100%;
+  margin-bottom: 5px;
+}
+
+.victory {
+  background-color: lightskyblue;
+}
+
+.defeat {
+  background-color: red;
 }
